refactor(crit): use utils file
This commit introduces `utils.go` for the helper functions to read the
magic value. It also adds some comments and bug fixes.
