+++ .gitignore
@@ -8,3 +8,4 @@ scripts/*.h
 scripts/expected.go
 scripts/output.go
 magic/magic.go
+crit/bin
