+++ Makefile
@@ -10,6 +10,8 @@ lint:
 build:
 	$(MAKE) -C scripts
 	$(GO) build -v ./...
+	# Build crit binary
+	$(MAKE) -C crit bin/crit
 
 test: build
 	$(MAKE) -C test
